was missing float on wait cmd

+++ MotionMeerkat/arguments.py
@@ -41,7 +41,7 @@ def arguments(self):
                                 self.parser.add_argument("--plotwatcher", help="Camera was a plotwatcher?",action="store_true",default=False)
                                 self.parser.add_argument("--frame_rate", help="frames per second",default=1)
 				self.parser.add_argument("--moglearning", help="Speed of MOG background detector, lowering values are more sensitive to movement",default=0.15,type=float)                                
-				self.parser.add_argument("--learning_wait", help="Wait X minutes to allow MOG to initialize?",default='0.5',type=int)                                		                                
+				self.parser.add_argument("--learning_wait", help="Wait X minutes to allow MOG to initialize?",default='0.5',type=float)                                		                                
                                 self.parser.add_argument("--subMethod", help="Accumulated Averaging [Acc] or Mixture of Gaussian [MOG] background method",default='MOG',type=str)                                
                                 self.parser.add_argument("--mogvariance", help="Variance in MOG to select background",default=16,type=int)                                
                                 self.parser.add_argument("--set_ROI", help="Set region of interest?",action='store_true',default=False)
